Added objectType, childNodes, textContent Fixed "attribute" wrt XML/list of XML distinction--needs more fixing Implemented __call() to help clarify (?) gen'd js code. Moved utility function to better organize jslib.js Fixed bug with __start that caused not all nodes to be deleted.    (length was being re-evaluated while deleting nodes) Revert to handling strings in __XML construction Implement __accum, for loops on client side Added debugging functions: error, dump, debugObj, debug
gensym now accepts a string and tacks its unique number onto the 
  end of that, so we can preserve variable names.
Removed Iter constructor in js IR--it's generated directly now
JS dictionary keys are now quoted
"Let" is generated as "Bind" in JS gen, to prevent side-effecting 
  code from being optimised out.
Fix beta-reduction of gen'd js code and turn it on.
Add a trivial test for this.