Typing for messages, according to the scheme we've been discussing recently.
Every function receives an extra parameter, which is passed through
applications, encoding the type of the messages that the process can
receive.  The parameter is fed back through spawn and included in the
type of the process id, ensuring that processes cannot be sent
messages of types that they are not expecting.

The type of process ids is no longer Int.  Instead, the id of a
process which can receive messages of type 't' is 'Mailbox t'.

Messages are no longer limited to variant types; a message can be any
Links value.

Sadly, since every function has an extra parameter, our types now look
rather odd. For instance, take has the type

    a -> Int -> a -> [b] -> [b]

It would be better to print the mailbox parameters differently, e.g.

    Int -> [b] -> [b] receives (a)

after to the Java/C++ throw syntax.

There's an infelicity in the system: mailbox types which include type
parameters are too general.  The value restriction would solve this,
but perhaps we can do something a bit better.

As a result of the change in type printing there are some failing
tests.  This is deliberate, really: they're a reminder that we need
some better print syntax for transformed types, and fast.