Generate show instances for recursive polymorphic types.
OCaml has an extension for recursive modules, but it doesn't allow
recursive functors (see

   A proposal for recursive modules in Objective Caml (Xavier Leroy;
   Version 1.1, May 13, 2003)
   http://caml.inria.fr/pub/papers/xleroy-recursive_modules-03.pdf)

We can do a pretty good impression of recursive functors, however:

   module rec R (P : S1) : S2 = B

   module R (P : S1) : S2 = 
     struct
       module rec C : S2 = B [R(P) -> Impl]
       include C
     end

A similar encoding works for mutually recursive functors, allowing a
fairly natural encoding of the show instances for mutually recursive
types.

(One assumption I've made is that every participant in a mutually
 recursive types must have the same number of parameters.  For
 example, the following is disallowed:

   type 'a f = ...
   and ('a,'b) g = ...

My reasoning is that for any set of mutually recursive types where
this does not hold one of the following must be true:

   1. The types are not truly mutually recursive
   2. The types are irregular, in which case we can't traverse them
      with O'Caml functions anyway [0]

Counterexamples are welcomed!

[0] Actually polymorphic recursion /is/ possible in O'Caml, contrary
    to what I'd previously thought.  See the above cited paper for
    details.

Also, fix a couple of bugs: allow parameterized recursive types that are not sum
or variant types; remove the executable with `make clean'.

(Aside: Is this a valid way to find the fixpoint of a binary type
 constructor (such as `type ('r,'a) seq = Nil | Cons of ('a * 'r)')?
 Is it any use?

  module type F = 
  sig
    type ('recur,'a) ofunct 
  end

  module type Fixed = 
  sig
    type 'a funct
  end

  module FixFunct (F : F) : Fixed =
  struct
    module rec C : Fixed =
    struct
      type 'a funct = ('a C.funct, 'a) F.ofunct 
    end
    include C
  end
)