Derived instances for rewrite.  The function process_children and the surrounding boilerplate are now generated automatically if there's a
     deriving (Rewriter)

annotation on the type declaration.

This doesn't handle much yet: no support for type aliases, polymorphic
variants, records and more.  It's sufficient for generating the
rewriters for Syntax.expression and Js.code, though.  It can't cope
with the instance for Types.datatype yet (no support for indirect type
recursion).

Any type constructors applied to the type on the rhs of its definition
must be instances of Functor.  For example

     type e = A | B of (e option list)

is ok, because option and list have Functor instances, but 

     type e = A | B of (e Stack.t)

is not because there's no Functor instance for Stack.t.

The code currently generated is a bit imperative, which makes me
nervous.  I can't see anything /obviously/ wrong with it, though.