OCaml implementation of QuickCheck.
Currently a bit cumbersome, mostly due to the need to simulate
dictionary synthesis by hand.  

It'll be interesting to see whether this will actually be useful.  It
seems like it should be possible to use it to check some properties of
(e.g.)  optimisers, serialisation, etc., but it's not clear how
difficult or worthwhile it'd be to actually write the tests.  It works
for checking properties of List.rev and function composition, anyway.