Add a sample testing framework.  For now, just the list tests.
Feedback / comments very welcome!

There's a bug because there's no reasonable way to compare types at
present.  I expect the right way is to have a canonical form for types
(minimised, then with type variables numbered according to some
order); for now just make the (wrong) assumption that equivalence of
types is the same as equivalence of printed representations.