Implemented Peter Thiemann's syntax for XML forests
  <#>...</#>

This makes it easy to include raw text at the beginning and end of XML
literals and will allow us to treat form bindings as XML.

Note: there's no reason why we have to stick with this syntax. I just
implemented it because it was easy to do.

I'm not sure if my changes to the lexer were entirely correct...

Is '#' supposed to be an opchar?

What are 'xml_opening' and 'xml_closing_tag' for (they don't appear to
be used)?