Fix the shift/reduce conflict introduced by r749.  (Yes, I'm fussy about grammar conflicts.)
The conflict is as follows.  In the following code

  <foo>{a : b -> c}</foo>

the spliced expression can be parsed as either

    (a : b) -> c

(i.e. form-bind expression 'a' of type 'b' to variable c) or as

    a : (b -> c)

(i.e. expression 'a' has type 'b -> c').  I've changed the grammar so
that type annotations aren't allowed on the expression on the lhs of
form bindings unless you put them in parentheses.  If you want the
first parse above you now actually have to write:

    (a : b) -> c