Add `with' syntax for records to make small changes more convenient
   (r with
      f1 = e1,
      ...
      fn = en)
  
The fields have to be present in the original record (although the
types associated with each field need not match what was there
previously: you can say ((x=1) with x="one")).  If you want to add
fields not already present you can use the existing syntax

   (f1 = e1,
    ...
    fn = en | r)

It'd be nice to unify these two syntaxes.  Perhaps something like a
'+' before a field name to indicate that it's not present in the
original record would be nice:

   (r with
      f1 = e1,
     +f2 = e2,
      ...
      fn = en)