Replaced Record_empty with Record_intro, which introduces an n-ary record. This immediately gives us much saner javascript code for constructing records.
It should be reasonably straightforward to generalise Record_extension
to extend a record with n fields, though the benefits will be less
noticeable (as far as I can tell none of our examples uses record
extension on the client!). Another obvious generalisation would be to
support record concatenation, but this may make type inference
difficult, so its probably best to leave that for now.

We should also think about improving Record_selection, but it isn't clear
what the best approach is.