Operators abs and app.
We have the following typing rules:

Standard abstraction results in a function that takes n arguments.

          G, x1:t1,...xn:tn |- e : t
  -------------------------------------------
  G |- fun (x1,...xn) { e } : (t1,...tn) -> t

The abs operator transforms a function of one argument, an n-tuple,
into a function of n arguments.

   G |- f : ((t1,...tn)) -> t
  -----------------------------
  G |- abs f : (t1,...tn) -> t

Standard application takes a function, f, of n arguments, and n
arguments e1 ... en, and applies f to e1 ... en.

      G |- f : (t1,...tn) -> t
   G |- e1 : t1, ... G |- en : tn
   -------------------------------
         G |- f(e1,...en) : t

The app operator takes a function, f, of n arguments, and an n-tuple
p, and applies f to p.

    G |- f : (t1,...tn) -> t
       G |- p : (t1,...tn)
    -------------------------
          G |- f app p : t

The syntax isn't great, but that's easily changed.  A more serious
concern: the typing rules require that the argument of the argument to
abs has tuple type, but we can't distinguish tuples from records in
our type language.  The result is that certain expressions receive a
type but fail at runtime (see next check-in for examples.)

The JS CPS code generated probably can and should be improved by
someone who has more of an idea what's going on.  It seems to work,
though, at least in simple cases.