Most of the main deriving bits completely rewritten now.  Also a new implementation of Show with much better style (pretty faint praise).
Confession: I did some object-oriented programming (and I liked it).
For an extensible fold over mutually recursive types it seems to offer
the right abstractions.

  1. Inheritance with selected overriding works better than passing
     around lots of arguments.

  2. Actual late binding works better than simulating it with dodgy
     open recursion schemes.

  3. The type signatures are much easier to understand as well.

I think it might be worth looking into adopting an OO fold for dealing
with expressions in Links.  For example, for our current AST ("IR" for
Sam) we'd have the signature
   
   class virtual ['a] fold :
     object
       method virtual crush : 'a list -> 'a

       method expression    : expression -> 'a
       method definition    : definition -> 'a
       method program       : program    -> 'a
     end

instead of this:

   val reduce_expression :
     ((expression -> 'b) -> expression -> 'b) -> 
     (expression * 'b list -> 'b) -> expression -> 'b
   val reduce_definition :
     ((expression -> 'b) -> expression -> 'b) -> 
     (expression * 'b list -> 'b) ->
     (definition * 'b list -> 'b) ->
     definition -> 'b
   val reduce_program :
     ((expression -> 'b) -> expression -> 'b) -> 
     (expression * 'b list -> 'b) ->
     (definition * 'b list -> 'b) ->
     ('b list * 'b -> 'b) ->
     program -> 'b
   
For Sam's proposed IR we'd have 

   class virtual ['a] fold :
     object
       method virtual crush    : 'a list -> 'a

       method value            : value            -> 'a
       method tail_computation : tail_computation -> 'a
       method binding          : binding          -> 'a
       method special          : special          -> 'a
       method computation      : computation      -> 'a
     end

instead of some monstrosity that I don't want to think about.  The
fold could be derived, of course.  (The new Camlp4 is supposed to have
fold generation built in, but it doesn't seem to work on the types I
wanted it for.)