Rationalize the shelve interface a bit more.
Add an extra arg to Repr.make to indicate the number of constructors.

Change the whizzy* names.

Remove allocate_id in favour of allocate; put it in a functor so that
it accepts fewer and better-typed arguments.

Do all allocation for record objects inside library code.  Change the
interface so that the recursion is clear: the function passed in to
Read.Utils.record now accepts an extra argument: the uninitialized
record object.  Remove the anomalous update_map.