New syntax for "overloaded function" application.
Previously you needed either a named type or lots of messing about
with functors for every method used for deriving.

For example, if you didn't have a declaration for the type 

   (bool * string) list option

in your program, but you had a value of that type that you wanted to
show you had to write either

   type made_up_name = (bool * string) list
                          deriving (Show)
   ...

      Show_made_up_name.show v

or 

   let module M = Show_list(Show_2(Show_bool)(Show_string)) in
       M.v

This change introduces some new "generics"-type syntax so that you can
write

   Show.show<(bool*string) list> v

instead.