Turn on sugar typing by default.
All tests now pass except for 

   Absence typing in variant patterns
   HasType pattern [3] 

which fail because we no longer have refinement typing and

   Incorrect message type sent to a process is a type error
   Receive types must unify (incorrect, closed rows)
   Receive types must unify (incorrect, open rows)
   Mailboxes are not polymorphic [2]

which are fussy about the exact format of error messages and

   Lexical scoping at top-level
   handle HTML with VAR -> HANDLER construction
   Case patterns (with redefined hd)
   continuation typing [3]
   Abs and app on non-tuple records
   Tables must have table type.
   Duplicate fields
   Labelization bug

which also fail in the trunk.

All programs in examples/ now pass type checking except for 

   extensible-case.links

which is probably something to do with refinement typing.