Start implementing an upcast operator.
  e : t <- s

means "coerce the expression e of type s to the type t". It's based on
Phil's coercion syntax in "well typed programs can't be blamed" (<=
didn't work because that's already used as the less-than-or-equal-to
operator, which is why I used <-).

Most of the functionality isn't implemented yet, but it works for some
basic types.