Reinstate r1395 with lots of fixes.
The frontend pipeline is in frontend.ml. Currently it only has four
stages, but we should pull out some more stages from sugar.ml. These
should include: desugaring of pages and formlets as well as pattern
matching compilation. Desugaring should happen on the AST for the
source language rather than that for the IR, and it shouldn't be mixed
with compilation to an IR as it is now.

The first pass in the frontend simply fills in the source_code buffers
for every position, which means we don't have to pass a 'lookup_pos'
function around all over the place like we used to.

Some silly bugs that were fixed in the new type checker:

  - A fold_left is used on a list of bindings in combination with
cons. Whenever this is done, the resulting list has to be reversed.
  - The first element of list literals was being dropped during
type-checking!