Fixed a bug where aliases weren't being expanded.
This required separating out concrete_types into two functions: one
for just expanding MetaTypeVars and another that in addition expands
aliases. Furthermore, the expand_aliases function had to be moved into
a separate module as it needs to call Instantiate.aliases which
depends on the Types module.

Also got rid of the blinking and random colours in type error
messages.