Fix a bug where we were not respecting the value restriction.
A simple example is

  sig f : (a) -> a
  var f = id(id)

which used to succeed, but now returns an error because the free rigid
type variable 'a' in the type of id(id) = (a) -> a cannot be
generalised, as id(id) is not a value.

The bug is fixed in the new type inference algorithm, but not in the
old one.

Other fixes:

 - got rid of `ForAll dodgeyness in unification and elsewhere
 - fixed subtle bugs in type-checking recursive functions