Rationalised date and time stuff.
We now have four functions:

  clientTime : () -> Int
  serverTime : () -> Int

  intToDate : (Int) -> (year:Int, month:Int, day:Int, hours:Int,
minutes:Int, seconds:Int)
  dateToInt : ((year:Int, month:Int, day:Int, hours:Int, minutes:Int,
seconds:Int)) -> Int

The first two functions return the number of seconds since 1970 from
the client and server respectively.

The other functions convert back and forth between the number of
seconds since 1970 and a record representation. The record
representation is based on that used by JavaScript.

The ranges of fields are as follows

  year (0-)
  month (0-11)
  day (1-31)
  hours (0-23)
  minutes (0-59)
  seconds (0-59)

The representation differs slightly from that of UNIX in that UNIX
years start from 1900 (i.e. in UNIX 1900 is represented in 0 whereas
in JavaScript and now Links 1900 is represented as 1900).