Implemented Ezra's rules for query compilation extended with order by and a limited form of limit / offset. These are integrated with the existing CPS interpreter using a non-standard direct-style interpreter. A nice feature is that the query compiler handles map, concatMap and sortBy, so we can compile away for comprehensions in the frontend using Phil and SPJ's comprehensive comprehensions desugaring rules.
I changed the db {...} syntax to query {...} as db is quite commonly
used as a variable name in our code. We might consider other
possibilities. Limit / offset are implemented as parameters to
query. The expression query [10] {...} denotes a query with limit 10
and offset 0. The expression query [10,20] {...} denotes a query with
limit 10 and offset 20.

All of the examples are now working with the new query
compiler. Currently only MySQL is supported and quite a lot of things
need tidying up. Of course, the effect-type system has not yet been
implemented, so we don't yet get the static guarantees we would like.