Put types into a normal form where there are no nested (type) expressions.  For example, the declaration
   type 'a list = Nil | Cons of 'a * 'a list

will be converted to something like  
 
   type 'a list = Nil | Cons of 'a t1
    and 'a t1 = 'a * 'a list

(applications of locally-bound constructors to type parameters are
considered atomic).

This makes the implementation simpler and significantly reduces the
size of generated code.  (I'm not absolutely sure it's the right thing
to do, though, because it appears to make user-defined specialization
more difficult.)

Still needs a bit of work.  Checking in now so that it doesn't get
lost.