More presence flag stuff. Presence type variables are now supported in the frontend. The needed row is now used when inferring types for updates.
In order to infer more useful types for inserts we will have to
augment the source syntax slightly (to include the labels of the
fields we intend to supply).

In order to compile updates sensibly we will have to do something
similar but simpler than what we now do for queries.