More fiddling with serialisation and the interpreter.
This commit gives an order of magnitude speed-up to the
interpreter. This is mainly due to the removal of a very expensive and
pointless operation that was happening every time the environment was
extended with a local environment: a global closure environment is
attached to each local environment and the union of the two closure
environments was being computed.

The interpreter is still slower than the old interpreter in the trunk
(about 5 or 6 times slower on my simple test case). I think that the
main reason for this is likely to be that the old interpreter has
separate global and local environments whereas the new interpreter has
a single environment. We should probably adapt the new interpreter to
use two environments as well.

I also experimented with representing the environment as a list
instead of a tree. This gave an order of magnitude slowdown. Results
may well differ if we had separate global and local environments.