"Fix show."
That's the effect; the details are that deriving has changed quite a bit.

The main change is that "type classes" are now simulated using records
rather than modules.  This means that there won't be any more
Undefined_recursive_module errors, for one thing.  For another, it
means that it's much easier to pass a "dictionary" around; before, to
write a "print" function that displays any showable value you'd have
to write

   module Print (Show : Show) =
   struct
      let print v = print_endline (Show.show v)
   end

   let module P = Print (Show_int) in P.print 3

whereas now you can write it as 

   let print show v = print_endline (Show.show v)

   print show_int 3

which seems like an improvement on the whole.

Other details:

The reworked deriving has a more liberal idea of normal form for type
declarations.  In particular, type constructor applications can appear
nested.  (This may be reverted soon, though.)

Top-level expressions generated by deriving are eta-expanded to avoid
the "let rec restriction".

Type representations (Typeable.TypeRep.t) now use lazy values rather
than thunks.

The number of functions in the "dictionary" is minimized; everything
else is an overloaded function, not a method.

Compilation of the whole tree is about 20% faster.

The special syntax for calling overloaded functions (not currently
used in the Links tree) is chnaged so as not to clash with built-in
syntax.

There's a new bug involving type representations: comparison of the
representations of certain recursive types does not terminate.  This
doesn't necessarily cause a problem in practice for Links, since only
the Pickle serialiser depends on this sort of thing and you can just
use Dump or Marshal instead.  Nevertheless, it's a bug and it needs to
be fixed, and will be.  There's a failing test case (currently
commented-out) in deriving-310/tests/inline.ml.