Improve support for regular expressions on the client.
Our regexp implementation is still rather dodgey, but at least this
should work more reliably than what we had before. It should cope with
everything except for 'Replace'. The current implementation relies on
the fact that regexps are encoded using the concrete syntax. In
particular, the encoding includes explicit 'Group' constructs.

Also, we're now outputting a correct syntax for grouping. Previously
we were using:

 (:? ... )

which happened to work because :? matches 0 colons! The correct syntax is:

 (?: ... )

which means 'group ... but don't remember the match'. I'm not quite
sure why we bother to do this anyway.

For consistency with the server we no longer wrap anchors around
regexps.

The server implementation is broken in that although composite ranges
are not lexed or parsed they are still treated as composite ranges:

 "a" =~ /^[a-zA-Z]$/           (1)

returns true, whereas

 "[a-zA-Z]" =~ /^[a-zA-Z]$/    (2)

returns false. Presumably this is because the server-side
pretty-printer for regexps isn't properly quoting strings. The client
correctly returns false for (1) and true for (2).

The tests in regex.js were already broken because they assume that
Links strings are represented as JS strings. They're now even more
broken.