properly append ts and vs components of the results of case-expressions when joining the results together.
it is not sufficient to simply do a union over all cases as for the
top-level plan. there might be an append on nested lists somewhere in
one of the case-expressions which generates new surrogate keys and
thereby destroys the property that the iter values of the different
cases are disjunct on all levels.

solution: simply append vs/ts the same way as in compile_list. since
this has to be done pair-wise, this means that on the top level for n
cases n-1 rownums are created for the new surrogate keys. however,
pathfinder seems to get rid of them quite nicely.