This is disabled by default, so it shouldn't break existing stuff. To enable it switch on the 'quantified_instantiation' setting.
Partial support for 'first-class' polymorphism. There are still some
issues that need addressing before this actually works properly. I
think type inference more-or-less works, but the generated IR is
definitely wrong in some cases. The difficulty is in generating the
correct type abstractions and applications (see the bug description in
typeSugar.ml).