if a case (switch) statement has no default case, don't SELECT the iterations for a case from the remaining iterations of its predecessors, but from the whole set.
this might be more efficient because fewer intermediate results have to be kept.
it might otherwise be a bit less efficient because all SELECTs operate on the
whole set of iterations.

at least for the current benchmarks, there is no significant difference in running
time.

At least the generated plans are a bit more compact.