* Experimental faster stitching code (deforests out intermediate data structures)  - some of this code is experimental and will go away
* Flat query benchmark (to evaluate overhead of shredding/sand on queries without any nesting)

Cherry-pick 70a02a9239203e64ac1f5c9c8586b6352a10f284

This is a huge mess in evalir.ml.
We should get rid of the slower stitching versions if possible.
There seems to be a bug in either (or both) of the versions. With
deforest_stitching on, there are results missing.