Treat unkinded type variables as being of kind Unl, not kind Any. Restore type signatures in prelude.links.
This has the advantage of requiring the smallest delta from existing
type signatures---if a signature worked in pre-linearity Links, it
should work now.  However, it does result in a somewhat odd behavior.
Consider the following definitions

> sig id1: (a)->a; fun id1(x){x}
> fun id2(x){x}

One might expect these to be equivalent; however, the first could not be
applied to a (non-End) channel, as the type variable 'a' is implicitly
restricted to kind 'Unl'.  In this case, the most general type signature
would be

> sig id: (a::Any)->a::Any; fun id(x){x}