Work in progress version of the Breakout clone I'm implementing.
I also ported a small coroutine library (still working on it) to make it work. But I ended up not exactly using the coroutines as I should. Once I sort out some problems that are still left and understand properly how it should work, I'll do it properly. For it's ok for testing.