This should hopefully make everything that matters up to date.
I moved the game examples to games folder and updated index.html accordingly.
New stuff since the last commit would be pacman and tetris as well as another version of the benchmark.
Also everything was adjusted to work with the latest Links if needed.
I had to change a few lines in irtojs.ml, because I changed _yield's signature, while optimizing it.