Module desugaring progress.
Absolute basics work now. It's possible to have modules, have them
desugared, and it'll rename everything correctly. As a concrete example,
the following works:

```
module A {
  var a = 0;
  var b = 0;

  module B {
    fun foo() {
      print("World!")
    }
    var a = 0;
  }

  fun foo() {
    print("Hello!");
    B.foo()
  }
}

A.foo()

```

Regarding deltas to the previous patch, this patch fixes a bug
previously with bindings getting reordered incorrectly (leading to
out-of-scope errors which were incorrect), and also handles renaming of
qualified variables correctly (for example, B.foo() inside A.foo()
correctly gets renamed to A.B.foo(), whereas it didn't before).