Intermediate commit.
Fix some errors with parsing, but for some reason it still won't parse
the open keyword.

Also work on the logic for imports -- in-module imports *should* (if
this code is correct) be included into the renaming path now.

Sketched out the chaser for external file imports, but that's not
implemented yet.