Fix projection out of provenance not working
(prov someField).3 for getting the row used to not work. There were two
reasons: First, type destructuring on Prov types was not implemented.
Second, we would only desugar the calls to `prov` and `data` later in
the rewriting pipeline. We do this in one step now, together with the
table/iteration stuff.