Begin reworking modules.
I'm hoping to implement the system from "A Theory of Name Resolution".
This requires several changes to the sugared AST, unique binder /
binding names, and a scope graph.

Currently I have the very start of the scope graph, and the
uniquifying pass.

Next up is to be more disciplined with qualified names, which need to be
parsed in as separate (unique) names. Then we can use the rather
beautiful algorithm to resolve them, as opposed to my ad-hoc nonsense.