Disable module chasing via qualified variable names
Modules should be explicitly imported using "open" instead of being
found using qualified variable names.

Chasing via qualified variable names results in problems when dealing
with partially-qualified names. In practice, many module systems require
either explicit import ordering (such as F#) or a set of top-level
imports (such as Haskell) -- so this isn't unprecedented.

It's probably possible to make the chaser smarter to take this into
account, but for now it's best to disable it, so we have a sensible
system for the release.