Newer version of modules system (#93)
This commit adds a new version of the modules system, based on the scope graph formalism of Neron et al ("A Theory of Name Resolution", ESOP 2015).

The module system (for now) is hidden behind a flag, "-m".

The implementation of the module system is entirely on the frontend, implemented as a desugaring pass. More details are on the Wiki.

Added files:

* uniquify.ml -- generates an AST with unique variable names (for both binding and bound occurrences)
* scopeGraph.ml -- specifies and constructs a "scope graph", a formalism for relating declarations and references, and for resolving references to declarations
* desugarModules.ml -- armed with a unique AST and scope graph, renames unique names to the correct fully-qualified names
* chaser.ml -- includes external files, allowing projects with multiple modules
* moduleUtils.ml -- helper functions for the above files

Additionally, there are unit tests which test the core functionality.