Always display fresh variables in error messages
This is a follow-up on my earlier fix to #43. During discusion
between Sam, Daniel and me we concluded that it is really
difficult to come up with a general and simple rule when
to hide fresh variables in error messages. It seems that the easiest
and safest thing to do is to always display all the variables in
error messages.