Rationalize (somewhat) the treatment of initial slashes in routes.
There are (effectively) two changes.  The first is that you no longer
need to provide the initial slash when specifying routes.  So, the
following two invocations of addRoute have the same effect:

    addRoute("/examples/", main)
    addRoute("examples/" main)

The second change is that you can now specify the behavior of the bare
server name without it being interpreted as the behavior for all URIs.
So, if you say:

     addRoute("", main)

then that will handle (e.g.) http://localhost:8080/, but not
http://localhost:8080/foo/bar.  However, if you actually *want* the old
behavior, you can still get it by specifying

    addRoute("/", main)