More testing and fixes to modules (#121)
Most examples seem to work now!

There is an issue with todoDb -- although the module desugaring pass
seems to do exactly what it's supposed to, a later pass seems to
rearrange the bindings in such a way that the 3 functions aren't treated
as mutually-recursive anymore. This is fixed (seemingly) when the
mutually-recursive "showList" function isn't called directly. More
investigation needed.

The other problem is that Pacman seems to cause a stack overflow.