Make PIDs gre-- abstract again
This commit fixes that infernal abstraction leak, also making them
random as opposed to sequential.

There is still a bit of work to do on jslib so that a similar thing is
done when creating processes client-side, especially if the client PID
is eventually serialised -- but that's for another day.

We have the best PIDs. Really, we do.