Mother of all refactorings!!
Json now no longer depends on Proc :)

This commit splits the JSONiser into two parts: a
possibly-side-effecting part in resolveJsonState, and a pure part in
json.ml.

Additionally, process serialisation is driven entirely by the map in
proc.ml as opposed to being extracted from the PIDs as before.

As a consequence of the JSON refactoring, the interface for websockets
is *much* nicer. This commit should also now ensure that processes and
event handlers are correctly serialised as a result of client calls,
too.

Grungy work, but I'm glad it's done!