Avoid creating cont. frame in request / accept
Before, a separate continuation frame was created to pass the session
channel when blocking in the evaluation of request / accept in
evalir.ml (according to @slindley, likely an artefact of copying the
pattern employed by the "recv" function).

This patch changes this, so that the value is returned directly,
avoiding creating the intermediate data structure.