Serialisable access points w/ realpages
This should allow access points to properly work with realpages.

A new `newAP` construct works a bit like spawnAt, taking a spawn
location as its argument. These are then spawned on the client.

It's not yet possible to do client-to-server or client-to-other-client
or server-to-client requests / accepts -- that is, AP request can't
cross boundaries yet. But soon!