Revamp JS implementation of session chans & APs
This commit brings the JavaScript implementation of session channels and
access points more in line with the server-side version. This will allow
us to more easily work with serialising endpoints when we get client <->
server AP communication going, as well as delegation eventually.

(also, the mandelcolor patch shows that we no longer have to hack around
to get client processes spawned properly. Hurrah!)