Remove old string_of_value code
We now use the pretty printer to print to a string buffer. We use a
custom formatter which ignores newlines and indentation.

Pretty printing is always disabled in web_mode.
Pretty printing is always enabled in the REPL.
Pretty printing is disabled by default for file/expression input but
can be enabled by setting print_pretty.
Colors are disabled by default, but can be enabled by setting
print_colors. We only print colors when pretty printing.