Revamp serveWebsockets()
This has been bugging me for a while. Now, the websocket connection URL
is a configuration option rather than an argument. Therefore, we only
need to do "serveWebsockets()" instead of, say,
"serveWebsockets('/ws/').

This patch therefore moves the connection URL out of requestData, with a
flag set in webserver.ml.