Add native readline functionality (fixes #51)
Uses the linenoise package and hooks everything together (which is a bit
of a faff with the Lexer).

This works (to my satisfaction) to be the same as rlwrap's version,
including histories, parsing of partial inputs, and handling of
interrupts such as ctrl-c and ctrl-d.

Other people might want to test to see if it is satisfactory for them
too.