Add ability to have "standard library" files
These are modules which are expected to be used in *some* modules, but
not all -- those should go in the prelude.

They are resolved as part of standard module resolution.

First example: factor out VDom, so that the junk doesn't need to be in
every file (hurrah!!!). You can now just do "open VDom" and you should
be good.

Also explicitly qualify StringHandler etc. in vdom.js accordingly.