[UNFINISHED] Work on recursive channel cancellation
The main issue is tying this together with trap -- which currently is
pure, but we need it to be in Lwt!

Maybe the best option is to lift some of the logic back out into evalir.
But to do that, we need the captured environments. Maybe a generalised
trap would be the way to go?