Begin to work on client-side session exceptions
Still a way to go yet. Variable inspection doesn't seem to be correct just yet alas. It might be we need to inspect further along the continuation stack (or something might just be bugged with the inspection algo. Who knows!)

Next up after that is needing to compile the send / receive operations specially, since they might invoke _SessionFail. We can't do the same approach as on the server, since we need to compile effect invocation statically instead of just swapping out the continuation.