Session Exceptions (again!) (#312)
* Session Exceptions

Turns out it's inherent to the design. Who'da thunk it.

* Fix @dhil's comments, compatibility with linked lists

* Update comment in channelVarUtils.ml

We can't use the IR visitor at present since it's parameterised by a
typing environment, which we don't have access to.

To add it would require a large refactor. But I think we could probably
use the tables in tables.ml to avoid this altogether, so I'd rather do
that (later on) than push a large refactor through that I'll end up not
needing.

* Fix choice compilation; add tests for offering

* compile receive normally if enable_exceptions not set

Additionally, add checks to ensure handlers flag set if exceptions flag
is set, and that exceptions flag is set if exceptions present in source.