Move manual install/remove steps to dune
Dune does not yet support dynamic install targets, I have replaced those
with the preliminary generation of the necessary install stanzas. These
are written in install.inc that is then included in the dune file. An
empty install.inc is kept in the root to make sure that the project
compiles even outside the opam installation environment.

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>