Use smart constructors in desugar* modules
This patch introduces more smart constructors and applies them in
several desugar* modules.  It also redefines implementation of a default
SugarConstructors.Make module to prevent creation of positions inside
the compiler pipeline.  The motivation for this is that positions
correspond to tokens in the source code and thus can only be
meaningfully constructed by the parser.

Fixes #421 and #423