Fixes #459
This is an attempt to fix #225 while allowing reuse of . for session types and qualified types. It is hacky and wrong, but passes all of the existing tests.

Explanation:

1. Create primary_datatype_loc nonterminal for primary_datatypes annotated with their location (so that their semantic actions have the same type as datatype)

2. Remove qualified_type_name from parenthesized_datatypes and make a new nonterminal qualified_or_primary_datatype that allows either a non-parenthesized primary_datatype or a qualified_or_primary_datatype.

3. Move TILDE datatype from session_type to primary_datatype and restrict the second component to be primary_datatype_loc.

The third change was the one I was least pleased about but seems necessary to parse existing type expressions using TILDE conventionally on variables, such as

# examples/sessions/forkbomb.links
```
mu a.!~a.a
```
Restricting what goes after the datatype is necessary so that a session_type cannot end with a tilde'd qualified type. That is, you still have to write
```
~(M.x)
```
but that is what we would have had to write before this change.