Explicit mutually-recursive functions and types (#528)
# Overview

Currently, `refineBindings.ml` does two things:

  1. Implicitly determines groups of mutually-recursive functions by building a call graph and topologically sorting it
  2. Hacks mutually-recursive types by inlining type aliases and inserting `mu` binders

While 1 might be reasonably sensible, 2 was a huge mistake, leading to massive slowdowns and incomprehensible error messages.

This patch implements the design we decided upon in #457, where mutually-recursive types and functions are explicitly defined in a `mutual` block. For example, consider the canonical example of odd / even numbers. This can be expressed as follows:

```
mutual {
  typename Even = [| Z | SuccE:Odd |];
  typename Odd = [| SuccO:Even |];

  sig three : () -> Odd
  fun three() {
    SuccO(SuccE(SuccO(Z)))
  }
}

three()
```

The key idea behind the patch is to adopt a degree of nominal typing when dealing with mutually-recursive types. Instead of the inlining pass, which would construct something like:

```
typename Even = mu X . [| Z | SuccE:[| SuccO:X |] |];
typename Odd = mu X . [| SuccO:[| Z | SuccE:[| SuccO:X |] |] |];
```

we now retain the type information, and compare recursive type applications directly. This has the advantage of making unification faster, as we are no longer comparing huge types, and also means that error messages no longer fill the screen with junk. Having explicit mutually-recursive functions and types is also useful when doing other passes (such as the new module system), and fixes some odd behaviour.

## Relation to issues
This patch addresses #457, and also fixes #455 and #412. It should also go some way to helping with #359, although there are other issues at play there which could probably be fixed by considering types as minimised automata.

# Technical design

## Frontend

Type alias constructor with `Type` is replaced with `Typenames`, which contains a list of mutually-recursive types. Types not defined in a `mutual` block are represented by a singleton list.

`mutual` blocks are parsed directly into `Typenames`, `Fun`, and `Funs` blocks.

## `refineBindings.ml`
Redundant: removed completely. Rejoice!

## `types.ml`
Each type group is associated with some metadata:
```
type tygroup = {
  id: int;
  type_map: ((quantifier list * typ) Utility.StringMap.t);
  linearity_map: bool Utility.StringMap.t
}
```
Here, `id` is a unique ID for the type groups; `type_map` stores data about each type in the group (in particular, the quantifiers and the desugared datatype); and `linearity_map` stores linearity information.

A new `tycon_spec` constructor, `Mutual of (quantifier list * tygroup ref)` is used when desugaring types in the same group. As we cannot count on having desugared datatypes for mutually-defined types since they can be cyclic, type applications to a mutually-defined type are compiled to a new semantic type `RecursiveApplication of rec_appl`, where `rec_appl` is defined as:

```
and rec_appl = {
  r_name: string;
  r_dual: bool;
  r_unique_name: string;
  r_args: type_arg list;
  r_unwind: type_arg list -> bool -> typ;
  r_linear: unit -> bool option
}
```

Here, `r_name` is the user-visible name, and `r_unique_name` is used internally for unification. `r_dual` is used to track whether the type application is dualised (in the context of session types) -- my treatment of dualisation follows that of Lindley & Morris (2016). `r_args` contains the type arguments. `r_unwind` is a function which takes a list of type arguments and a boolean stating whether the unwinding should be dualised, and returns the one-step unfolding of the type: this is used
during unification. Finally, `r_linear` is a function which returns `None` if linearity has not yet been calculated; `Some true` if the recursive application is linear; and `Some false` if not.

## `desugarDatatypes.ml`

Constructs the semantic types for mutually-recursive syntactic types by creating a new `tygroup`; populating the alias environment with `Mutual` `tycon_spec`s for types in the same mutual blocks; and then desugaring each type. After this, it becomes possible to calculate the linearity of each type by calculating recursive applications; topologically-sorting the SCCs; and propagating linearity through the application graph. Finally, desugaring of the remaining program is calculated with an
updated `alias_env` with each of the types promoted to a "proper" `Alias`.

## `desugarMutual.ml`
A simple pass which desugars `mutual` blocks into `Typenames` and `Funs` blocks.

## `unify.ml`
Modified to track recursive applications as well as mu-bound function names. When unifying two `RecursiveApplication`s, checks to see whether their (unique) internal name is equal.
If so, unification succeeds.
If a recursive application is already witnessed, unification succeeds. Unification with a primitive fails. Otherwise, unwinds the recursive type, and records that this has been done in order to avoid nontermination.

## `instantiate.ml`
Some duplicate code refactored out; also some functionality exposed which allows us to instantiate recursive aliases.

## Testing
Test cases:
  * Lists (basic recursive types and functions on them)
  * Odd/even numbers (mutually-recursive types)
  * Only functions and typenames may be in a recursive block
  * Unguarded recursive types cannot unify with primitives
  * Type variables are not shared in a mutual block
  * Linearity: Linearity is propagated correctly, but also unrestricted types can still be treated as unrestricted
  * Duplicate bindings disallowed