Some error message easy wins (#536)
The result of two train-rides' work refactoring how we do error messages in the frontend.

This firstly fixes #531 by removing the logic duplication in formatting HTML -- in turn, this should make it less painful to add new errors (and encourage people to do that, instead of using failwith).

This secondly addresses #514 by renaming the confusingly-named SugarError to DesugaringError and attaching a variant describing a stage to it. In doing this, two things came up:

In some places, SugarError was used where an InternalError was more appropriate
In desugarLAttributes, dummy positions were used, so I've refactored it to use proper position information now
There are also some other minor refactorings to make things nicer.