Keep closure information for functions where IR type-checking fails.
The IR type-checker is designed such that it can recover from encountering
an ill-typed binding by adding the expected type (found in the type
annotation) to the environment and checking the rest of the program.
However, in such a case, we didn't add the closure information  before,
which could lead to unnecessary follow-up errors.