Resolves issue #576.
The "smart" constructor `tuple' has a rather hacky method for
preserving 1-tuples, which is not always correct. In particular, it
introduced a soundness regression for formlets.

This commit augments the interface of `tuple' with an optional boolean
flag to enable or disable the said hack. By default the hack is on.