Sound types for the `addRoute'-family of functions.
This patch refines the type of unsafeAddRoute, and its relatives, in
particular `addRoute', which previously had the type

   addRoute : (String, (String) ~> Page) ~> ()

which elaborates to

   addRoute : (String, (String) ~e~> Page) ~f~> ()

where `e' and `f' are distinct effect variables. Consequently the
invariant that all user-definable effectful operations are handled
prior to reaching the toplevel is broken, for instance

   addRoute("/", fun(_) {do Fail})

would cause a "no handler installed for Fail" error to occur at
runtime when visiting the root page, even though one might have
wrapped the `addRoute' in a handler, i.e.

   handle(addRoute("/", fun(_) {do Fail})) {
     case Fail -> ()
   }

This does not handle the effects of the anonymous function, because it
gets executed in a separate context where the handler is not
installed.

To ensure that all user-definable computational effects are handled,
this patch "corrects" the type of `unsafeAddRoute' and its associated
functions, as a result the type of `addRoute' is now

   addRoute : (String, (String) {hear{_}}~> Page) ~> ()

which elaborates to

   addRoute : (String, (String) {hear{p}}~> Page) ~f~> ()

In other words, a page generating function must handle all of its
effects -- save for the two built-in effects `hear' and `wild'.