Desugar modules properly (#602)
* Various improvements to the module desugaring pass.

This commit fixes various bugs related to modules:

- It is no longer possible to inadvertently peek across module
  boundaries.

- The previous implementation would prefix structural names, which is
  incorrect. Now structural names are... structural.

- Modules work in the REPL.

Other improvements:

- Structured code with annotated with reasoning and justifications.

- The implementation desugars modules in a single pass rather than
  multiple passes.

* Removes nonsensical qualification of structural data constructors.

The previous desugaring pass for modules would quailify data
constructor labels which is entirely nonsensical given that Links is
predominately structural typed. Qualifying data constructor labels
inadvertently break the structural typing of polymorphic variants.

It would be reasonable to have monomorphic variants along side our
polymorphic variants as in OCaml. However whether we want that is the
subject of another discussion.

* Allow modules to be empty.

* Various improvements to module desugaring including better error reporting.

* Improve printing of module members.

Ensures that the implementation-specific details of name resolution
does not leak. Module members are printed as

    A.B.C.x

rather than

    A$0$B$0$C$0$x

Module members or typenames that may escape their scope gets prefixed
with a question mark, for example

    fun () { typename T = Int; 3 : T }()

The result is printed as `3 : ?.T`.

Fixes #456.

* Addresses @SimonJF's comments.

* Alien blocks must be desugared after the chaser has run.

The chaser may inline additional alien blocks, hence it is crucial
that the desugaring pass for alien blocks is run after the chaser.

Fixes #292.

* Fixes pretty printing of type constructor names in recursive
datatypes.

* CP expressions introduce a new scope.

* Explain the resolution strategy.

* Make the translation of qualified names in JavaScript slightly nicer.

Previously the JavaScript compiler would translate a qualified name
like "Breakout.main" to

      _B_r_e_a_k_o_u_t_fullstop_m_a_i_n_2539(__kappa)

now the compiler generates names that are easier to read and closer to
their original source representation

      _Breakout_0$main_2539(__kappa)

* @dhil: Import scope tests from the 'proper-modules' branch.

Original message:
@frank-emrich: added tests for type name scoping in modules

* Update module test suite.

* Adds a few tests for import.